Authorization with gates: where attendes or the owner can delete the attenteed from the events

+++ app/Http/Controllers/Api/AttendeeController.php
@@ -8,6 +8,7 @@ use App\Http\Traits\CanLoadRelationships;
 use App\Models\Attendee;
 use App\Models\Event;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class AttendeeController extends Controller
 {
@@ -46,6 +47,7 @@ class AttendeeController extends Controller
      */
     public function show(Event $event,Attendee $attendee)
     {
+
         return new AttendeeResource(
             $this->loadRelationships($attendee)
         );
@@ -64,6 +66,7 @@ class AttendeeController extends Controller
      */
     public function destroy(Event $event,Attendee $attendee)
     {
+        Gate::authorize('delete', [$event,$attendee]);
         $attendee->delete();
 
         return response(status: 204);
